Move errors validation from the mutation to the createUser integration test
